requirements.txt
requests
google-cloud-storage
apache-beam[gcp]
apache-airflow[google]
apache-airflow-providers-google
apache-airflow-providers-apache-beam
